requirements.txt
streamlit
pandas
openai
scikit-learn
numpy
python-dotenv